refactor(application_utils): move all constants into module
