+++ crates/utils/common/Cargo.toml
@@ -5,3 +5,7 @@ edition = "2021"
 
 [dependencies]
 chrono = { workspace = true }
+compile-time = "=0.2.0"
+const-str = { workspace = true }
+dotenvy_macro = { workspace = true }
+reqwest = { workspace = true }
